Break up cabal project and module structure.
