+++ chat-bots-contrib/chat-bots-contrib.cabal
@@ -1,0 +1,125 @@
+cabal-version: 3.0
+name:          chat-bots-contrib
+version:       0.1.0.0
+synopsis:      Bot behavior and Server modules for the chat-bots library.
+bug-reports:   https://github.com/cofree-coffee/cofree-bot
+license:       MIT
+author:        Solomon, Asad, and the Cofree-Coffee community
+maintainer:    ssbothwell@gmail.com
+category:      bot, chat-bot, finite-state-machines, matrix
+
+--------------------------------------------------------------------------------
+
+common common-settings
+  default-language:   Haskell2010
+  default-extensions:
+    DataKinds
+    DeriveFunctor
+    DeriveGeneric
+    DerivingVia
+    FlexibleContexts
+    FlexibleInstances
+    GeneralizedNewtypeDeriving
+    ImportQualifiedPost
+    InstanceSigs
+    LambdaCase
+    NamedFieldPuns
+    OverloadedStrings
+    PatternSynonyms
+    RecordWildCards
+    ScopedTypeVariables
+    StandaloneKindSignatures
+    TupleSections
+    TypeApplications
+    TypeFamilies
+    TypeOperators
+
+  ghc-options:
+    -Wall
+    -Wcpp-undef
+    -Widentities
+    -Wincomplete-record-updates
+    -Wincomplete-uni-patterns
+    -Wno-name-shadowing
+    -Wpartial-fields
+    -Werror=missing-home-modules
+
+--------------------------------------------------------------------------------
+
+common common-libraries
+  build-depends:
+    , base           >=2 && <5
+    , bytestring
+    , filepath
+    , matrix-client
+    , network-uri
+    , profunctors
+    , text
+    , these
+    , typed-process
+
+--------------------------------------------------------------------------------
+
+library
+  import:
+    , common-libraries
+    , common-settings
+
+  hs-source-dirs:  src
+  exposed-modules:
+    Data.Chat.Bot.Calculator
+    Data.Chat.Bot.Calculator.Language
+    Data.Chat.Bot.CoinFlip
+    Data.Chat.Bot.GHCI
+    Data.Chat.Bot.Hello
+    Data.Chat.Bot.Jitsi
+    Data.Chat.Bot.Jitsi.Dictionary
+    Data.Chat.Bot.Magic8Ball
+    Data.Chat.Bot.Updog
+    Data.Chat.Server.Matrix
+
+  build-depends:
+    , aeson
+    , attoparsec
+    , containers
+    , chat-bots
+    , data-fix
+    , directory
+    , hint
+    , http-client
+    , lens
+    , monad-loops
+    , mtl
+    , pretty-simple
+    , process
+    , random
+    , transformers
+    , vector
+    , xdg-basedir
+
+--------------------------------------------------------------------------------
+
+test-suite chat-bots-contrib-test
+  import:
+    , common-libraries
+    , common-settings
+
+  type: exitcode-stdio-1.0
+  hs-source-dirs: test
+  main-is: Spec.hs
+
+  build-depends:
+    , attoparsec
+    , chat-bots
+    , chat-bots-contrib
+    , data-fix
+    , hspec
+    , hspec-core
+    , lens
+    , mtl
+    , parsec
+    , template-haskell
+
+  other-modules:
+    Scripts
+    TestServer
