+++ chat-bots-contrib/test/Scripts.hs
