+++ chat-bots/chat-bots.cabal
@@ -1,0 +1,93 @@
+cabal-version: 3.0
+name:          chat-bots
+version:       0.1.0.0
+synopsis:      A library for building bots compositionally. 
+bug-reports:   https://github.com/cofree-coffee/cofree-bot
+license:       MIT
+author:        Solomon, Asad, and the Cofree-Coffee community
+maintainer:    ssbothwell@gmail.com
+category:      bot, chat-bot, finite-state-machines, matrix
+
+--------------------------------------------------------------------------------
+
+common common-settings
+  default-language:   Haskell2010
+  default-extensions:
+    DataKinds
+    DeriveFunctor
+    DeriveGeneric
+    DerivingVia
+    FlexibleContexts
+    FlexibleInstances
+    GeneralizedNewtypeDeriving
+    ImportQualifiedPost
+    InstanceSigs
+    LambdaCase
+    NamedFieldPuns
+    OverloadedStrings
+    PatternSynonyms
+    RecordWildCards
+    ScopedTypeVariables
+    StandaloneKindSignatures
+    TupleSections
+    TypeApplications
+    TypeFamilies
+    TypeOperators
+
+  ghc-options:
+    -Wall
+    -Wcpp-undef
+    -Widentities
+    -Wincomplete-record-updates
+    -Wincomplete-uni-patterns
+    -Wpartial-fields
+    -Werror=missing-home-modules
+
+--------------------------------------------------------------------------------
+
+common common-libraries
+  build-depends:
+    , base           >=2 && <5
+    , bytestring
+    , matrix-client
+    , network-uri
+    , profunctors
+    , text
+    , these
+    , typed-process
+
+--------------------------------------------------------------------------------
+
+library
+  import:
+    , common-libraries
+    , common-settings
+
+  hs-source-dirs:  src
+  exposed-modules:
+    Data.Chat.Bot
+    Data.Chat.Bot.Monoidal
+    Data.Chat.Bot.Context
+    Data.Chat.Server
+    Data.Chat.Server.Repl
+    Data.Chat.Utils
+    Control.Monad.ListT
+    Parsing
+
+  build-depends:
+    , aeson
+    , attoparsec
+    , containers
+    , data-fix
+    , directory
+    , filepath
+    , hint
+    , http-client
+    , lens
+    , monad-loops
+    , mtl
+    , pretty-simple
+    , process
+    , random
+    , transformers
+    , vector
