--- cofree-bot.cabal
@@ -1,153 +1,0 @@
-cabal-version: 3.0
-name:          cofree-bot
-version:       0.1.0.0
-synopsis:      A library for building bots compositionally. 
-bug-reports:   https://github.com/cofree-coffee/cofree-bot
-license:       MIT
-author:        Solomon, Asad, and the Cofree-Coffee community
-maintainer:    ssbothwell@gmail.com
-category:      bot, chat-bot, finite-state-machines, matrix
-
---------------------------------------------------------------------------------
-
-common common-settings
-  default-language:   Haskell2010
-  default-extensions:
-    DataKinds
-    DeriveFunctor
-    DeriveGeneric
-    DerivingVia
-    FlexibleContexts
-    FlexibleInstances
-    GeneralizedNewtypeDeriving
-    ImportQualifiedPost
-    InstanceSigs
-    LambdaCase
-    NamedFieldPuns
-    OverloadedStrings
-    PatternSynonyms
-    RecordWildCards
-    ScopedTypeVariables
-    StandaloneKindSignatures
-    TupleSections
-    TypeApplications
-    TypeFamilies
-    TypeOperators
-
-  ghc-options:
-    -Wall
-    -Wcpp-undef
-    -Widentities
-    -Wincomplete-record-updates
-    -Wincomplete-uni-patterns
-    -Wno-name-shadowing
-    -Wpartial-fields
-    -Werror=missing-home-modules
-
---------------------------------------------------------------------------------
-
-common common-libraries
-  build-depends:
-    , base           >=2 && <5
-    , bytestring
-    , filepath
-    , matrix-client
-    , network-uri
-    , profunctors
-    , text
-    , these
-    , typed-process
-
-executable cofree-bot
-  import:
-    , common-libraries
-    , common-settings
-
-  main-is:        Main.hs
-  hs-source-dirs: app
-  build-depends:
-    , aeson
-    , barbies
-    , cofree-bot
-    , directory
-    , filepath
-    , mtl
-    , optparse-applicative
-    , xdg-basedir
-    , yaml
-
-  other-modules:
-    , Options
-    , Options.Env
-    , Options.Config
-    , Options.Types
-    , Options.Parser
-
---------------------------------------------------------------------------------
-
-library
-  import:
-    , common-libraries
-    , common-settings
-
-  hs-source-dirs:  src
-  exposed-modules:
-    CofreeBot
-    CofreeBot.Bot
-    CofreeBot.Bot.Behaviors
-    CofreeBot.Bot.Behaviors.Calculator
-    CofreeBot.Bot.Behaviors.Calculator.Language
-    CofreeBot.Bot.Behaviors.CoinFlip
-    CofreeBot.Bot.Behaviors.GHCI
-    CofreeBot.Bot.Behaviors.Hello
-    CofreeBot.Bot.Behaviors.Jitsi
-    CofreeBot.Bot.Behaviors.Jitsi.Dictionary
-    CofreeBot.Bot.Behaviors.Magic8Ball
-    CofreeBot.Bot.Behaviors.Updog
-    CofreeBot.Bot.Context
-    CofreeBot.Utils
-    CofreeBot.Utils.ListT
-    Parsing
-
-  build-depends:
-    , aeson
-    , attoparsec
-    , containers
-    , data-fix
-    , directory
-    , hint
-    , http-client
-    , lens
-    , monad-loops
-    , mtl
-    , pretty-simple
-    , process
-    , random
-    , vector
-    , xdg-basedir
-
---------------------------------------------------------------------------------
-
-test-suite cofree-bot-test
-  import:
-    , common-libraries
-    , common-settings
-
-  type: exitcode-stdio-1.0
-  hs-source-dirs: test
-  main-is: Spec.hs
-
-  build-depends:
-    , attoparsec
-    , cofree-bot
-    , data-fix
-    , hspec
-    , hspec-core
-    , lens
-    , mtl
-    , parsec
-    , template-haskell
-
-  other-modules:
-    Scripts
-    TestServer
