--- src/CofreeBot.hs
@@ -1,12 +1,0 @@
-module CofreeBot
-  ( module Behaviors,
-    module Bot,
-    module Context,
-    module Utils,
-  )
-where
-
-import CofreeBot.Bot as Bot
-import CofreeBot.Bot.Behaviors as Behaviors
-import CofreeBot.Bot.Context as Context
-import CofreeBot.Utils as Utils
