--- src/CofreeBot/Bot/Behaviors.hs
@@ -1,18 +1,0 @@
-module CofreeBot.Bot.Behaviors
-  ( module Calculator,
-    module CoinFlip,
-    module GHCI,
-    module Hello,
-    module Jitsi,
-    module Magic8Ball,
-    module Updog,
-  )
-where
-
-import CofreeBot.Bot.Behaviors.Calculator as Calculator
-import CofreeBot.Bot.Behaviors.CoinFlip as CoinFlip
-import CofreeBot.Bot.Behaviors.GHCI as GHCI
-import CofreeBot.Bot.Behaviors.Hello as Hello
-import CofreeBot.Bot.Behaviors.Jitsi as Jitsi
-import CofreeBot.Bot.Behaviors.Magic8Ball as Magic8Ball
-import CofreeBot.Bot.Behaviors.Updog as Updog
